Renamed 'viewset_mixin.py' to 'viewset_mixins.py' to be compliant with DRF
